# Changed Linear to Ring topology for attention and MLP all-gather