--- a/models/demos/gemma3/tt/gemma_image_attention.py
+++ b/models/demos/gemma3/tt/gemma_image_attention.py
@@ -373,7 +373,7 @@ class TtGemmaImageAttention(LightweightModule):
                 dim=3,
                 multi_device_global_semaphore=self.tt_ccl.get_and_cycle_ag_semaphore_handles(),
                 num_links=1,
-                topology=ttnn.Topology.Linear,
+                topology=ttnn.Topology.Ring,
                 barrier_semaphore=self.tt_ccl.get_and_cycle_barrier_semaphore_handle(),
                 chunks_per_sync=10,
                 num_workers_per_link=2,