--- a/models/demos/gemma3/tt/gemma_image_mlp.py
+++ b/models/demos/gemma3/tt/gemma_image_mlp.py
@@ -118,7 +118,7 @@ class TtGemmaImageFeedForward(LightweightModule):
                 dim=1,
                 multi_device_global_semaphore=self.tt_ccl.get_and_cycle_ag_semaphore_handles(),
                 num_links=1,
-                topology=ttnn.Topology.Linear,
+                topology=ttnn.Topology.Ring,
                 barrier_semaphore=self.tt_ccl.get_and_cycle_barrier_semaphore_handle(),
                 chunks_per_sync=10,
                 num_workers_per_link=2,