### Ticket
(https://github.com/tenstorrent/tt-metal/issues/28519)

### Problem description
All_gather_async has previously used Linear topology in all 3
all_gather operation in Gemma3 (attention, pre-MLP and MLP),
which was bad for performance.

### What's changed

**Attention**:

- Utilization of ideal time for all_gather increased from 43.65% to 65.65%.
- Device time has been improved from 1009.3us to 671us measured in test_vision_attention.

**MLP**:

- Utilization of ideal time for all_gather increased from 45.9% to 68.06%.
- Device time has been improved from 23208.7us to 15652us measured in test_vision_mlp.

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/runs/18159771271/job/51688436608) CI passes
- [ ] (For models and ops writers) [T3000 demo tests](https://github.com/tenstorrent/tt-metal/actions/runs/18260825485) CI passes (if applicable, since this is required for release)
- [ ] [(Single-card) Demo tests (with perf runs) ](https://github.com/tenstorrent/tt-metal/actions/runs/18159765260)